微修正

+++ router/items.go
@@ -45,7 +45,7 @@ func GetItem(c echo.Context) error {
 	if err != nil {
 		return c.JSON(http.StatusBadRequest, err)
 	}
-	item, err := model.GetItemByID(itemID)
+	item, err := model.GetItemByID(uint(itemID))
 	if err != nil {
 		return c.JSON(http.StatusNotFound, err)
 	}
@@ -65,7 +65,7 @@ func PostOwners(c echo.Context) error {
 	if err != nil {
 		return c.JSON(http.StatusBadRequest, err)
 	}
-	user, err := model.GetUserByID(body.UserID)
+	user, err := model.GetUserByID(uint(body.UserID))
 	if err != nil {
 		return c.JSON(http.StatusBadRequest, err)
 	}
@@ -73,7 +73,7 @@ func PostOwners(c echo.Context) error {
 	if err != nil {
 		return c.JSON(http.StatusForbidden, err)
 	}
-	item, err := model.GetItemByID(itemID)
+	item, err := model.GetItemByID(uint(itemID))
 	if err != nil {
 		return c.JSON(http.StatusBadRequest, err)
 	}
